requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-SocketIO
openai
psycopg2-binary
python-dotenv
eventlet
google-genai